Extract PlanNodeDecorrelator
