+++ presto-main/src/main/java/com/facebook/presto/sql/planner/optimizations/ScalarAggregationToJoinRewriter.java
@@ -14,49 +14,37 @@
 package com.facebook.presto.sql.planner.optimizations;
 
 import com.facebook.presto.metadata.FunctionRegistry;
-import com.facebook.presto.metadata.Signature;
 import com.facebook.presto.spi.type.BigintType;
 import com.facebook.presto.spi.type.BooleanType;
 import com.facebook.presto.spi.type.TypeSignature;
-import com.facebook.presto.sql.ExpressionUtils;
-import com.facebook.presto.sql.planner.DependencyExtractor;
 import com.facebook.presto.sql.planner.PlanNodeIdAllocator;
 import com.facebook.presto.sql.planner.Symbol;
 import com.facebook.presto.sql.planner.SymbolAllocator;
 import com.facebook.presto.sql.planner.iterative.Lookup;
+import com.facebook.presto.sql.planner.optimizations.PlanNodeDecorrelator.DecorrelatedNode;
 import com.facebook.presto.sql.planner.plan.AggregationNode;
 import com.facebook.presto.sql.planner.plan.AggregationNode.Aggregation;
 import com.facebook.presto.sql.planner.plan.AssignUniqueId;
 import com.facebook.presto.sql.planner.plan.Assignments;
 import com.facebook.presto.sql.planner.plan.EnforceSingleRowNode;
-import com.facebook.presto.sql.planner.plan.FilterNode;
 import com.facebook.presto.sql.planner.plan.JoinNode;
 import com.facebook.presto.sql.planner.plan.LateralJoinNode;
 import com.facebook.presto.sql.planner.plan.PlanNode;
 import com.facebook.presto.sql.planner.plan.ProjectNode;
-import com.facebook.presto.sql.planner.plan.SimplePlanRewriter;
-import com.facebook.presto.sql.tree.DefaultTraversalVisitor;
 import com.facebook.presto.sql.tree.Expression;
 import com.facebook.presto.sql.tree.FunctionCall;
-import com.facebook.presto.sql.tree.LogicalBinaryExpression;
 import com.facebook.presto.sql.tree.QualifiedName;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.Iterables;
 
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.function.Predicate;
-import java.util.stream.Collectors;
 
 import static com.facebook.presto.sql.analyzer.TypeSignatureProvider.fromTypeSignatures;
 import static com.facebook.presto.sql.planner.optimizations.PlanNodeSearcher.searchFrom;
-import static com.facebook.presto.sql.planner.optimizations.Predicates.isInstanceOfAny;
-import static com.facebook.presto.sql.planner.plan.SimplePlanRewriter.rewriteWith;
 import static com.facebook.presto.sql.tree.BooleanLiteral.TRUE_LITERAL;
 import static com.google.common.collect.ImmutableList.toImmutableList;
 import static java.util.Objects.requireNonNull;
@@ -70,6 +58,7 @@ public class ScalarAggregationToJoinRewriter
     private final SymbolAllocator symbolAllocator;
     private final PlanNodeIdAllocator idAllocator;
     private final Lookup lookup;
+    private final PlanNodeDecorrelator planNodeDecorrelator;
 
     public ScalarAggregationToJoinRewriter(FunctionRegistry functionRegistry, SymbolAllocator symbolAllocator, PlanNodeIdAllocator idAllocator, Lookup lookup)
     {
@@ -77,12 +66,13 @@ public class ScalarAggregationToJoinRewriter
         this.symbolAllocator = requireNonNull(symbolAllocator, "symbolAllocator is null");
         this.idAllocator = requireNonNull(idAllocator, "idAllocator is null");
         this.lookup = requireNonNull(lookup, "lookup is null");
+        this.planNodeDecorrelator = new PlanNodeDecorrelator(idAllocator, lookup);
     }
 
     public PlanNode rewriteScalarAggregation(LateralJoinNode lateralJoinNode, AggregationNode aggregation)
     {
         List<Symbol> correlation = lateralJoinNode.getCorrelation();
-        Optional<DecorrelatedNode> source = decorrelateFilters(lookup.resolve(aggregation.getSource()), correlation);
+        Optional<DecorrelatedNode> source = planNodeDecorrelator.decorrelateFilters(lookup.resolve(aggregation.getSource()), correlation);
         if (!source.isPresent()) {
             return lateralJoinNode;
         }
@@ -185,7 +175,6 @@ public class ScalarAggregationToJoinRewriter
             Symbol nonNullableAggregationSourceSymbol)
     {
         ImmutableMap.Builder<Symbol, Aggregation> aggregations = ImmutableMap.builder();
-        ImmutableMap.Builder<Symbol, Signature> functions = ImmutableMap.builder();
         for (Map.Entry<Symbol, Aggregation> entry : scalarAggregation.getAggregations().entrySet()) {
             FunctionCall call = entry.getValue().getCall();
             Symbol symbol = entry.getKey();
@@ -215,156 +204,5 @@ public class ScalarAggregationToJoinRewriter
                 scalarAggregation.getStep(),
                 scalarAggregation.getHashSymbol(),
                 Optional.empty()));
-    }
-
-    private Optional<DecorrelatedNode> decorrelateFilters(PlanNode node, List<Symbol> correlation)
-    {
-        PlanNodeSearcher filterNodeSearcher = searchFrom(node, lookup)
-                .where(FilterNode.class::isInstance)
-                .skipOnlyWhen(isInstanceOfAny(ProjectNode.class));
-        List<FilterNode> filterNodes = filterNodeSearcher.findAll();
-
-        if (filterNodes.isEmpty()) {
-            return decorrelatedNode(ImmutableList.of(), node, correlation);
-        }
-
-        if (filterNodes.size() > 1) {
-            return Optional.empty();
-        }
-
-        FilterNode filterNode = filterNodes.get(0);
-        Expression predicate = filterNode.getPredicate();
-
-        if (!isSupportedPredicate(predicate)) {
-            return Optional.empty();
-        }
-
-        if (!DependencyExtractor.extractUnique(predicate).containsAll(correlation)) {
-            return Optional.empty();
-        }
-
-        Map<Boolean, List<Expression>> predicates = ExpressionUtils.extractConjuncts(predicate).stream()
-                .collect(Collectors.partitioningBy(isUsingPredicate(correlation)));
-        List<Expression> correlatedPredicates = ImmutableList.copyOf(predicates.get(true));
-        List<Expression> uncorrelatedPredicates = ImmutableList.copyOf(predicates.get(false));
-
-        node = updateFilterNode(filterNodeSearcher, uncorrelatedPredicates);
-        node = ensureJoinSymbolsAreReturned(node, correlatedPredicates);
-
-        return decorrelatedNode(correlatedPredicates, node, correlation);
-    }
-
-    private Optional<DecorrelatedNode> decorrelatedNode(
-            List<Expression> correlatedPredicates,
-            PlanNode node,
-            List<Symbol> correlation)
-    {
-        Set<Symbol> uniqueSymbols = DependencyExtractor.extractUnique(node, lookup);
-        if (uniqueSymbols.stream().anyMatch(correlation::contains)) {
-            // node is still correlated ; /
-            return Optional.empty();
-        }
-        return Optional.of(new DecorrelatedNode(correlatedPredicates, node));
-    }
-
-    private static Predicate<Expression> isUsingPredicate(List<Symbol> symbols)
-    {
-        return expression -> symbols.stream().anyMatch(DependencyExtractor.extractUnique(expression)::contains);
-    }
-
-    private PlanNode updateFilterNode(PlanNodeSearcher filterNodeSearcher, List<Expression> newPredicates)
-    {
-        if (newPredicates.isEmpty()) {
-            return filterNodeSearcher.removeAll();
-        }
-        FilterNode oldFilterNode = Iterables.getOnlyElement(filterNodeSearcher.findAll());
-        FilterNode newFilterNode = new FilterNode(
-                idAllocator.getNextId(),
-                oldFilterNode.getSource(),
-                ExpressionUtils.combineConjuncts(newPredicates));
-        return filterNodeSearcher.replaceAll(newFilterNode);
-    }
-
-    private PlanNode ensureJoinSymbolsAreReturned(PlanNode scalarAggregationSource, List<Expression> joinPredicate)
-    {
-        Set<Symbol> joinExpressionSymbols = DependencyExtractor.extractUnique(joinPredicate);
-        ExtendProjectionRewriter extendProjectionRewriter = new ExtendProjectionRewriter(
-                idAllocator,
-                joinExpressionSymbols);
-        return rewriteWith(extendProjectionRewriter, scalarAggregationSource);
-    }
-
-    private static boolean isSupportedPredicate(Expression predicate)
-    {
-        AtomicBoolean isSupported = new AtomicBoolean(true);
-        new DefaultTraversalVisitor<Void, AtomicBoolean>()
-        {
-            @Override
-            protected Void visitLogicalBinaryExpression(LogicalBinaryExpression node, AtomicBoolean context)
-            {
-                if (node.getType() != LogicalBinaryExpression.Type.AND) {
-                    context.set(false);
-                }
-                return null;
-            }
-        }.process(predicate, isSupported);
-        return isSupported.get();
-    }
-
-    private static class DecorrelatedNode
-    {
-        private final List<Expression> correlatedPredicates;
-        private final PlanNode node;
-
-        public DecorrelatedNode(List<Expression> correlatedPredicates, PlanNode node)
-        {
-            requireNonNull(correlatedPredicates, "correlatedPredicates is null");
-            this.correlatedPredicates = ImmutableList.copyOf(correlatedPredicates);
-            this.node = requireNonNull(node, "node is null");
-        }
-
-        public Optional<Expression> getCorrelatedPredicates()
-        {
-            if (correlatedPredicates.isEmpty()) {
-                return Optional.empty();
-            }
-            return Optional.of(ExpressionUtils.and(correlatedPredicates));
-        }
-
-        public PlanNode getNode()
-        {
-            return node;
-        }
-    }
-
-    private static class ExtendProjectionRewriter
-            extends SimplePlanRewriter<PlanNode>
-    {
-        private final PlanNodeIdAllocator idAllocator;
-        private final Set<Symbol> symbols;
-
-        ExtendProjectionRewriter(PlanNodeIdAllocator idAllocator, Set<Symbol> symbols)
-        {
-            this.idAllocator = requireNonNull(idAllocator, "idAllocator is null");
-            this.symbols = requireNonNull(symbols, "symbols is null");
-        }
-
-        @Override
-        public PlanNode visitProject(ProjectNode node, RewriteContext<PlanNode> context)
-        {
-            ProjectNode rewrittenNode = (ProjectNode) context.defaultRewrite(node, context.get());
-
-            List<Symbol> symbolsToAdd = symbols.stream()
-                    .filter(rewrittenNode.getSource().getOutputSymbols()::contains)
-                    .filter(symbol -> !rewrittenNode.getOutputSymbols().contains(symbol))
-                    .collect(toImmutableList());
-
-            Assignments assignments = Assignments.builder()
-                    .putAll(rewrittenNode.getAssignments())
-                    .putAll(Assignments.identity(symbolsToAdd))
-                    .build();
-
-            return new ProjectNode(idAllocator.getNextId(), rewrittenNode.getSource(), assignments);
-        }
     }
 }
